refacto(configuration): share YAML loader and settings in common/
API and Playground duplicated config schemas and the YAML/${ENV} loader.
Move the shared pieces to a root common package so both apps extend one source of truth.
